+++ src/main/java/rocks/massi/data/joins/UsersGroupsRepository.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface UsersGroupsRepository extends JpaRepository<UsersGroups, UsersGroups.UsersGroupsKey> {
     List<UsersGroups> findByUserId(String userId);
 
-    List<UsersGroups> findByGroupId(Integer groupId);
+    List<UsersGroups> findByGroupId(Long groupId);
 
     @Transactional
-    void deleteByGroupId(Integer groupId);
+    void deleteByGroupId(Long groupId);
 }
